Android build commit.  Refactor and clean up so that node can be checked out to AOSP's external/ directory and then can be compiled simply with "make node".  I had to make a stub Android.mk file in the root directory and put the real binary building one under android/ as I couldn't get the BUILD_EXECUTABLE call to coexist nicely with the call all-subdir-makefiles.
The AOSP version of v8 is also older that the one in node v0.10
(missing a few things required) and it was being compiled as a
static library anyways so I figured we might as well switch to
the version packed up inside node.  I pulled over the  makefiles
from external/v8 and found that there seems to be a lot of variable
clobbering between the two.  I didn't trace down the exact root
cause but to work around it I just renamed some of the variables
like GEN1, GEN2, etc. to NODE_GEN1, NODE_GEN2, etc.  For some
reason they were taking on the values defined by the external/v8
makefiles.

Other important changes are to remove the -fvisibility=hidden flag
on v8, which seems to cause problems when I try and load
cross-compiled NPM modules in node.  Also standardize on the
platform+arch name to be "android-arch" as it is a bit weird to
have "android-arch-arm" which is the default with AOSP's v8
makefiles.